00-SampleCodes/Working with Load Balancing and Auto-Scaling/requirements.txt
-e .
aws-cdk.core>=1.3.0
aws-cdk.aws_ec2

aws-cdk.aws-autoscaling
aws-cdk.aws_elasticloadbalancingv2
aws_cdk.aws_rds
